backend/requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dotenv==0.9.9
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.57.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pandas==2.1.2
pillow==11.1.0
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
